requirements.txt
flask
sqlalchemy
flask_sqlalchemy
load_dotenv
requests
pytest
google-genai
